requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
requests
python-dotenv
